Make script to get github names.
A number of other things happened in this commit. A basic shell
environement was created with shell.nix. Also a Makefile with several
commands was fleshed out. A script to extract names from the
stackoverflow files was created. In total, the structure to get common
users between github and stackoverflow is in place.